Update README.md
The link to download the PowerShell helper goes to the GitHub page for that file, not a raw version. This is confusing for someone who's just trying to get the helper to install - they're just as likely as not to right-click and download from the link. The proposed change adds a second "raw file" link next to the helper link to make it clear the first link is not the raw script and to provide a way to download the raw script from the readme.